refactor: rename error_formatting.py -> formatting.py
Shorter, clearer module name for the error/warning formatting helpers.
(errors.py is intentionally avoided — that name should be reserved for
exception types, not formatters.)

  omnibenchmark/error_formatting.py     -> omnibenchmark/formatting.py     (format_yaml_warning)
  omnibenchmark/cli/error_formatting.py -> omnibenchmark/cli/formatting.py (pretty_print_parse_error)

Importers and package docstrings updated; ARCHITECTURE.md regenerated.
