+++ omnibenchmark/cli/dashboard.py
@@ -11,7 +11,7 @@ import click
 from omnibenchmark.core import BenchmarkExecution
 from omnibenchmark.core.dashboard import create_bettr_dashboard
 from omnibenchmark.logging import logger
-from omnibenchmark.cli.error_formatting import pretty_print_parse_error
+from omnibenchmark.cli.formatting import pretty_print_parse_error
 from omnibenchmark.model.validation import BenchmarkParseError
 from .debug import add_debug_option
